Create planets data

+++ rocket.h
@@ -9,8 +9,8 @@ class Rocket : public SolarObject
 {
     Q_OBJECT
 public:
-    explicit Rocket(Planet* startPlanet, unsigned int v,
-                    unsigned int angle, QObject *parent = 0);
+    explicit Rocket(Planet* startPlanet, double v,
+                    double angle, QObject *parent = 0);
 
 signals:
 
@@ -20,7 +20,7 @@ private:
     /**
      * @brief ROCKET_MASS rocket mass in kilograms
      */
-    static const unsigned int ROCKET_MASS = 2000000;
+    static const double ROCKET_MASS = 2000000;
 };
 
 #endif // ROCKET_H
